requirement.txt
google-generativeai
PyPDF2
python-dotenv
requests
beautifulsoup4
fastapi
uvicorn[standard]
python-multipart
streamlit